vm: Make vm_fault_quick_hold_pages take a pointer
Rather than passing a virtual address (usually cast from a pointer),
pass a pointer to vm_fault_quick_hold_pages as the starting address.
With CHERI this allows the bounds and permission to be checked before
mapping the memory in question for use by the caller.

Effort:		CHERI upstreaming
Sponsored by:	Innovate UK