sendfile(2): rework compat implementation
Rename the sendfile() function to kern_sendfile() and extend it with two
function pointer arguments to copy in the header/trailer structure and
the create uio's for the header and trailer as required.  Use this
to allow the removal of freebsd32_do_sendfile() which was a near idential
duplicate of sendmail().

While here, restyle calls to kern_sendmail to use a compound literal.

Effort:		CHERI upstreaming
Sponsored by:	Innovate UK