mmap: Add CHERI PROT_ flags
PROT_CAP, when combined with PROT_READ and/or PROT_WRITE causes mappings
to support loading or storing CHERI capabilities (subject to a number
of restrictions based on the mapping request).

If either of PROT_CAP or PROT_NO_CAP are set, then the value of
the PROT_CAP flag bit defines the page protections and capability
permissions for a given mapping.

In the upcoming implementation, PROT_CAP maps to VM_PROT_READ_CAP and
VM_PROT_WRITE_CAP depending on the values of PROT_READ and PROT_WRITE.
PROT_NO_CAP maps to VM_PROT_NO_IMPLY_CAP to avoid add _CAP permissions
when explicitly requested not to or to allow supeset tests to succeed on
max_protection when reducing permissions.

Effort:		CHERI upstreaming
Sponsored by:	Innovate UK, DARPA, AFRL